moved web ui logic to web tool/core
