+++ src/blenderbim/blenderbim/core/search.py
@@ -1,6 +1,2 @@
 def show_scene_elements(spatial):
     spatial.show_scene_objects()
-
-
-def open_web_ui(search):
-    search.open_web_ui()
