+++ src/blenderbim/blenderbim/core/web.py
@@ -1,0 +1,2 @@
+def open_web_ui(web):
+    web.open_web_ui()
